# Conflicts:
#	app/Http/Controllers/HomeController.php
#	resources/js/components/customer/Header.vue
#	resources/js/components/customer/home/ListProduct.vue
#	resources/js/router.js
#	resources/js/store.js
#	routes/api.php
#	routes/web.php

+++ database/migrations/2020_02_28_160609_create_products_table.php
@@ -19,8 +19,8 @@ class CreateProductsTable extends Migration
             $table->text('description');
             $table->text('slug')->unique();
             $table->integer('quantity');
-            $table->decimal('price');
-            $table->decimal('promotion');
+            $table->double('price');
+            $table->double('promotion');
             $table->integer('idCategory')->unsigned();
             $table->foreign('idCategory')->references('id')->on('category');
             $table->integer('idProductType')->unsigned();
